Remove relations() from trade profit calculations
Trade's main use is it to bolster diplomatic relationships between
empires - and the diplomatic bonus is based on the amount of trade
profit made. If the relationship is far too damaged (or is bad to
begin with) to make profit this leaves few ways to turn the
relationship around. Removing the relations rating from the
calculation opens the door for unlikely allies or reparations to be
made. This helps to make the diplomatic game more dynamic (although,
there are other problems with the system that undercut this overall
design).