Merge GalaxyShap generate methods
The two methods would only diverge in what they did when the galaxy
had more than 5000 stars - which is no longer supported - making the
two methods identical.

In GalaxyFactory the shape.generate() call is preserved because this
class should not assume what a UI component is doing as it's outside
of it's domain. The call to generate() does not pose any meaningful
performance impact.