Correct missile base defence stats
I couldn't find any documentation of how this worked in the original
game (and was too lazy to do some statistical analysis) but adding
your offense statistic (computer level) to your defence statistic
(missile and beam defence) seems like an obvious mistake. Even if it
isn't, it is counter intuitive and doesn't really add to the gameplay
in a meaningful way - so it is removed.

The beam defence is assumed to be fixed to 1 as that is what is left
if you remove ECM from bomb defence (which I am pretty sure is
correct) and missile defence is the same as bomb defence.

This does generally remove power from missile bases when they were
already maybe not the most effective to begin with... Their power may
be revisited at a later date.