Somewhat standardise alliance "chance"
It's not really a chance but a threshold at which things happen.
"Chance" mirrors the trade helper function nicely, however, and will
likely be reworked soon to be more chance related anyway.

Also, this could use further standardising as some factors are not
rolled into the helper function. This will be revisited later.