Remove Xilmi's makeDiplomaticOffers()
As Xilmi's diplomat AI is taking more and more code from the base AI
superclass (with the view to eventually merge the two AIs)
makeDiplomaticOffers() no longer aligned with what Xilmi’s AI would
accept if proposed to it. A quick read through the method did not show
anything that I wanted to copy over to the base AI so the method is
simply removed.