Merge Orion, enemy and player star types
The generation for all three star types is now the same. The extra
complexity in the old enemy and player homeworld star type generation
didn't feel impactful so it is now removed. This does not preclude
reintroduction of similar mechanics in the future but the current
implementation was deemed not worth supporting.