Move dialogueContactType() to embassy
This is not a perfect fit as embassy isn’t really about localizing
strings for user display but this allows Leader to become closer to a
POJO which seems like the correct design for Leader. It will probably
need to be moved again in the future.

This also the first use of a Java 8+ feature. Sticking to Java 8 isn’t
a goal and the project will generally be assumed to require the latest
JDK.