Move ANCHOR_SYSTEM into a field
sessionVar() just stores and retrieves values into a static HashMap.
This is in pretty much, in every way, a worse version of just storing
it in a static field (better discoverability, less error prone, more
efficient, type safe, etc). In the long run, sessionVar() should just
be removed but it is used in a lot of places and that will take time.
So for now only this use of it is being removed.