Remove threaten attack
Conceptually, it doesn't really make sense to threaten someone who is
already attacking you. If you are attacking someone - the possibility
that they might attack back is probably the first one you would
consider - so what is the outcome that you can hold as leverage when
making the threat?

Practically, this made very little difference to how the AI behaved or
the options given to the player.

This change does break how warnings to the player are handled but this
will be reworked in later commits.