Remove player threaten spy activity
To use this optimally would require a lot of tedious micromanagement.
In the non-optimal case this doesn't seem to provide any interesting
decisions making. Therefore, the option is simply removed.

The AI will still warn against spy activity. This is because nobody
cares if the AI is having fun and it creates interesting dynamics when
an empire can be falsely accused of spying.