Rework AI warn logic
The majority of the business logic changes are in AIDiplomat.

This commit changes the AI warn logic so that if any warning is issued
the second warning means war - where previously this would only happen
if it was the same type of incident and it was before the cool down
period ended.

This also includes a lot of incidental clean up and some bug fixes
which did not get committed separately because fixing something that
was going to be re-written anyway is a waste of time (i.e. apologies
for a messy commit).