Makes all wars immediate
Fundamentally, war is about communicating intent to the player and the
AI (i.e. it has no mechanical effect). As such, delaying the
signalling only worsens the player experience and makes it harder to
learn how the AI thinks. To this end "war footing" is removed and all
wars are declared as they are deemed necessary.