Don't automatically trigger war for allies
If AtWarWithAllyIncident always triggers war then AI is always dragged
into wars of their allies straight away. This limits nuance in the
diplomatic game so is removed.

Your relationship with the AI, which has an ally that you are at war
with, will still deteriorate due to this incident over time.