Fix Alert message
Previously, the Alert message would assume the transport was destroyed
by the empire controlling the colony. This isn't always correct (i.e.
a friendly colony might be orbited by a hostile fleet). The message
now blames the destruction of the transport on one of the empires that
has done the transports damage.

Long term this is probably not a useful part of the notification since
it's hard to assign just a single empire to blame in practise and
should be removed but due to licencing restrictions this is currently
not done.