Remove singleton instance
In GalaxyFactory static access to the singleton glues the business
logic to a single implementation making the code harder to reuse.
Therefore, replace static access to field access (this value could be
injected at construction time later).

In StarSystemFactory singleton accessor implies that it is a stateless
object - but this is only true because of static access it has to
helper classes. This will be replaced in the future so static
singleton is removed in favour of old-fashioned constructor use.