Defer declaration of war
Previous work to make war declaration immediate introduced a bug which
would make the game hang indefinitely when war was started due to
bombing of an enemy planet. This is now addressed by deferring war
declaration to a certain part of turn processing.

Ideally, stability of the program wouldn't depend so much on the
context of when certain calls are made but addressing that was deemed
too much work for now.