Remove erroneous oath breaker forgiveness
I can't think of why declaring a war of opportunity (or a war of hate,
though unlikely) on your ally should not count as breaking your oath.
I will, of course, eat humble pie if turns out that this is key in
some way - but I can't make sense of it at the moment.