Reimplement immediate war
It seems like the only thing that previously necessitated deferring
declaring war was Diplomat.noticeIncident() method which, previously,
cased the game to freeze. However, further investigation revealed that
this method is completely redundant as war is declared based on
incidents in Diplomat.makeDiplomaticOffers() anyway. This allows us to
return to the simpler and more responsive immediate war declaration
implementation (as well as allows us to simplify the code by removing
redundant code).