Remove init() methods
Deferring class initialising to another method in the constructor
doesn't provide much in terms of clarity (especially, if all it is
called is "init()") and makes it harder (/impossible) for the compiler
to tell if fields are final - to this end the majority of them have
been removed. There might still be more such init() methods hiding in
the code base but they were not obvious enough to be picked up in this
clean up sweep.