Remove maneuverability from CombatEntity
Maneuverability only impacts your ability to retreat if the ship is
hit by Technological Nullifier special. This behaviour would not be
obvious to player and the added complexity doesn't generate
interesting interactions. As such maneuverability is removed and
replaced with maxMove for the purposes of retreating (which is more
intuitive to the player).