Move isForgotten() to where it is used
Conceptually, it is out of DiplomaticIncident's scope to know if it is
forgotten or not so this logic is a better fit to code that makes this
decision. DiplomaticEmbassy is maybe not quite the right place to
decide if the incident is still relevant to the player or not but
redesigning this is out of scope for this change as this is already an
improvement and, therefore, worth committing.