Remove lazy initialisation of static fields
In these two cases of lazy initialisation addressed in this commit the
classes are relatively small in terms of data stored in them. This,
combined with the fact that they are referenced in static fields,
means that the benefit of delaying initialisation is extremely small
but comes with a cost of not being able to declare fields final and
general added complexity when reading the code. As such the trade off
was deemed not worth it and lazy initialisation has been removed.