Remove Xilmi's canDeclareWar() implementation
Xilimi's and base canDeclareWar() implementation differ by
inShipRange() in place of inEconomicRange(). This effectively means
that Xilimi's AI could declare war while only the active party was in
range of the recieving party's colonies (base AI required for both to
be in range). While this asymmetric diplomacy might be interesting
Xilimi's implementation is not consistent with how the rest of the AI
works and it is removed for that reason.

In the future inShipRange() might replace inEconomicRange() in AI
logic but it will be done in one commit across the board (or as close
to as I can manage).