Stop ExpansionIncident from encouraging genocide
If the number of systems an empire may have without penalty increases
when players become inactive there is an incentive to kill off
players. However, the game is worse when the player fights to kill off
a player. So this incentive is removed by considering number of
players at the start of the game only.