Change ExpansionIncident to impact late game only
Previously, ExpansionIncident would penalise large planet count
empires as soon as they have more than 5 planets. This was because it
tracked the total number of colonised planets - not total planets in
the galaxy.

This has the effect of exasperating poor starts as you can't
compensate for planet quality with quantity as easily. It also feels
off as empires seem to care about your development long before they
make contact.

For these two reasons the incident has been reworked to have no impact
on the early game and kick in only when an empire has taken over fixed
percentage of the galaxy (given the number of players in the galaxy).

Play tests suggest that this results in a better game feel.