Remove special handling of some techs for trading
This encodes the assumption that some techs are more valuable, despite
their research cost, than others (i.e. that there is a balance issue
with techs). This is undesired because when the balance of the techs
is adjusted (a thing I will eventually do) this code will break and
will be a pain to maintain.

i.e. Rather than maintaining the assumption that there is an imbalance
it would be more productive to address the imbalance and this code is
one of the things that makes addressing the imbalance harder.